adding doamin layer

+++ src/comments/application/comments.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Comment, CommentDocument } from '../domain/entity/comments.schema';
 import { CreateCommentDTO, UpdateCommentDTO } from './dto/comments.dto';
@@ -13,24 +17,37 @@ export class CommentsService {
     private CommentModel: Model<CommentDocument>,
     protected commentsRepository: CommentsRepository,
   ) {}
+  // domain factory
+  createComment(createCommentDTO: CreateCommentDTO) {
+    const contentLength = createCommentDTO.content.length;
+    if (contentLength < 3 || contentLength > 100) {
+      // throw new BadRequestException('length error');
+    }
+
+    const newCommentModel = new this.CommentModel(createCommentDTO);
+    return newCommentModel;
+  }
+
   async createCommentForSelectedPost(
     createCommentDTO: CreateCommentDTO,
-  ): Promise<Comment> {
-    const newComment = new this.CommentModel(createCommentDTO);
+  ): Promise<CommentDocument> {
+    const newComment = this.createComment(createCommentDTO);
     await this.commentsRepository.save(newComment);
     return newComment;
   }
+
   async updateComment(
     id: ObjectId,
     updateCommentDTO: UpdateCommentDTO,
   ): Promise<boolean> {
     const comment = await this.commentsRepository.findCommentById(id);
-    comment.content = updateCommentDTO.content;
+    comment.setContent(updateCommentDTO.content);
     const isCommentUpdated = this.commentsRepository.save(comment);
     return isCommentUpdated;
   }
   async deleteComment(id: ObjectId): Promise<boolean> {
     const comment = await this.commentsRepository.findCommentById(id);
+    if (!comment) throw new NotFoundException('not found');
     const isCommentDeleted = this.commentsRepository.delete(comment);
     return isCommentDeleted;
   }
